Created all CRUD functions

+++ view/admin/listposts.php
@@ -4,14 +4,14 @@
 
 <?php foreach ($posts as $post) : ?>
 
-	<h2> <?= $post->getTitle(); ?> </h2>
+	<h2><a href="../public/index.php?action=readpost&id=<?= $post->getID();?>"> <?= $post->getTitle(); ?> </a> </h2>
 	<p> <?= $post->getContent(); ?> </p>
 	<p>
-		<a class="btn btn-success" href="../public/index.php?action=updatepost&id=<?= $post->getID();?>"> Modifier l'article </a>
-		<a class="btn btn-danger" href="../../public/index.php?action=deletepost"> Supprimer l'article </a>
+		<a class="btn btn-success" href="../public/index.php?action=updatepost&id=<?= $post->getID();?>" onclick="return confirm('Êtes-vous sûr de vouloir modifier cet article?')"> Modifier l'article </a>
+		<a class="btn btn-danger" href="../public/index.php?action=deletepost&id=<?= $post->getID(); ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet article?')"> Supprimer l'article </a>
 	</p>
 
 <?php endforeach; ?>
 
 <?php $content = ob_get_clean(); ?>
-<?php require("../public/templates/admin/admin_template.php");
+<?php require("../view/templates/admin/admin_template.php");
